backend/requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
pyjwt
httpx
cryptography
google-cloud-bigquery
google-cloud-secret-manager
langchain
langgraph
anthropic
numpy
scipy
